app/core/itr_middleware.py
import time
from starlette.middleware.base import BaseHTTPMiddleware
from app.core.logger import get_logger

logger = get_logger("middleware")

class ITRMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request, call_next):
        start_time = time.time()
        response = await call_next(request)
        duration = round(time.time() - start_time, 3)
        logger.info(f"{request.method} {request.url.path} completed in {duration}s")
        return response
