requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
celery[redis]
playwright
